Synchronize on simpleExoPlayer when updating listener (#1677)
# Why

Fixes https://github.com/expo/universe/issues/1657.

**tldr**
`NullPointerException` was being raised because `mSimpleExoPlayer` was `null` when invoking `getExtraStatusFields`.

# How

I wrapped `callStatusUpdateListener()` with a synchronized block (on `mSimpleExoPlayer`) so it isn't released while we may still need it.

# Test Plan

BuildKite will test it for us, I can't reproduce this behavior on a real device.

fbshipit-source-id: 5c75f40