Make SecureStore return null for missing keys consistently across platforms, fix tests
On iOS we'd return undefined (`resolve(nil)`) for missing keys. The docs and Android return null, and this commit makes that consistent.

Added tests to test-suite to catch this. Also fixed the tests in test-suite that were bogus and throwing errors that were caught and causing the tests to pass for the wrong reasons.

Found a discrepancy between Android and iOS for duplicate keys in different iOS keychain services or different Android aliases -- for now the behavior will be different on each platform (iOS isolates them, Android doesn't and will throw if you try to use Alias B to decode a value stored with Alias A).

Fixes https://github.com/expo/expo/issues/983

fbshipit-source-id: 7af4669